+++ components/LikeButton/LikeButton.jsx
@@ -7,7 +7,7 @@ import HeartUnselected from '../../assets/graphics/favorite-unselected.svg';
 import styles from './LikeButton.module.css';
 
 export default function LikeButton({ dateObj }) {
-  const dateId = dateObj.id;
+  const dateId = dateObj.id.toString();
   const dateName = dateObj.name;
   const store = Store.useStore();
   const likedDates = store.get('likedDates');
@@ -25,7 +25,7 @@ export default function LikeButton({ dateObj }) {
       action: isFavorite ? 'Unliked Date' : 'Liked Date',
       label: dateName,
     });
-    localStorage.setItem('likedDates', newLikedDates);
+    localStorage.setItem('likedDates', JSON.stringify(newLikedDates));
     store.set('likedDates')(newLikedDates);
   };