+++ scenes/Search/Results/Results.jsx
@@ -1,24 +1,10 @@
 import React, { useState } from 'react';
-import ReactGA from 'react-ga';
 import { Checkbox } from '@material-ui/core';
-import { Experiment, Variant } from 'react-optimize';
 
 import styles from './Results.module.css';
-import Select from '../../../components/Select/Select';
 import DateCard from '../../../components/DateCard/DateCard';
 import Store from '../../../store';
-import { useFilters, filterDates, dateSorterNewest, dateSorterOldest } from '../../../utils';
-
-const options = [
-  {
-    value: 'Newest',
-    label: 'Newest',
-  },
-  {
-    value: 'Oldest',
-    label: 'Oldest',
-  },
-];
+import { useFilters, filterDates, dateSorterNewest } from '../../../utils';
 
 export default function Results() {
   const store = Store.useStore();
@@ -27,24 +13,13 @@ export default function Results() {
   const isFilterBarExpanded = store.get('isFilterBarExpanded');
   const [filters] = useFilters();
 
-  const [sortBy, setSortBy] = useState(options[0]);
   const [likedOnly, setLikedOnly] = useState(false);
-  let filteredDates = filterDates(dateObjs, filters).sort(
-    sortBy.value === 'Oldest' ? dateSorterOldest : dateSorterNewest,
-  );
+  let filteredDates = filterDates(dateObjs, filters).sort(dateSorterNewest);
   if (likedOnly) {
-    filteredDates = filteredDates.filter(dateObj => likedDates.includes(dateObj.id));
+    filteredDates = filteredDates.filter(dateObj => likedDates.includes(dateObj.id.toString()));
   }
   const resultsLength = filteredDates.length;
 
-  const toggleSort = val => {
-    ReactGA.event({
-      category: 'Interaction',
-      action: 'Select Sort',
-      label: val.value,
-    });
-    setSortBy(val);
-  };
   return (
     <div>
       {!isFilterBarExpanded && (
@@ -53,29 +28,15 @@ export default function Results() {
             {resultsLength} result{resultsLength !== 1 && 's'}
           </p>
           <div className={styles.sortRowSpacer}></div>
-          <Experiment id="xYIuiUCYTeGwXxODjB4B3Q">
-            <Variant id="0">
-              <div className={styles.selectContainer}>
-                <Select
-                  value={sortBy}
-                  size="small"
-                  onChange={toggleSort}
-                  options={options}
-                ></Select>
-              </div>
-            </Variant>
-            <Variant id="1">
-              <span className={styles.sortBy}>Only Favorites:</span>
-              <Checkbox
-                color="primary"
-                checked={likedOnly}
-                onChange={e => {
-                  setLikedOnly(e.target.checked);
-                }}
-                inputProps={{ 'aria-label': 'primary checkbox' }}
-              />
-            </Variant>
-          </Experiment>
+          <span className={styles.sortBy}>Only Favorites:</span>
+          <Checkbox
+            color="primary"
+            checked={likedOnly}
+            onChange={e => {
+              setLikedOnly(e.target.checked);
+            }}
+            inputProps={{ 'aria-label': 'primary checkbox' }}
+          />
         </div>
       )}
       <div className={styles.resultList}>